converted the sqlite package to TypeScript
Moved sqlites.js to a .ts file and added typings.
The module.exports function is now replaced by an init function.

The sqlite package has two tsconfig files:
 - tsconfig.ref.json will be used for TypeScript project references
 - tsconfig.json can be used for a standalone build of the package

The top level tsconfig file references "tsconfig.ref.json" and future
TypeScript packages should be added here too. This way
"npm run tsc" can be executed to build all packages,
and a "go to Definition" is possible with supported editors.

The "exports" field of package.json publishes everything from "dist",
existing require() paths do not have to change.

jest has no support for exports but it works with a "moduleNameMapper".
See server and saltcorn-data package.json
