+++ .github/workflows/nodejs.yml
@@ -46,6 +46,7 @@ jobs:
       - run: lerna bootstrap --no-ci
         env:
           PUPPETEER_SKIP_CHROMIUM_DOWNLOAD: "true"
+      - run: npm run tsc
       - run: pip3 install requests pytest
       - run: echo '127.0.0.1 example.com sub.example.com sub1.example.com sub2.example.com sub3.example.com sub4.example.com sub5.example.com' | sudo tee -a /etc/hosts
       - run: echo '127.0.0.1 otherexample.com' | sudo tee -a /etc/hosts
