+++ packages/sqlite/package.json
@@ -4,11 +4,23 @@
   "description": "Sqlite structures for Saltcorn, open-source no-code platform",
   "homepage": "https://saltcorn.com",
   "scripts": {
-    "test": "echo \"Error: no test specified\""
+    "test": "echo \"Error: no test specified\"",
+    "tsc": "tsc -p tsconfig.json",
+    "clean": "rm -rf ./dist/*"
   },
   "author": "Tom Nielsen",
   "license": "MIT",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "exports": {
+    "./*": "./dist/*.js"
+  },
+  "typesVersions": {
+    "*": {
+      "*": [
+        "dist/*"
+      ]
+    }
+  },
   "dependencies": {
     "@saltcorn/db-common": "0.6.1-beta.3",
     "sqlite3": "^5.0.2"
@@ -16,5 +28,10 @@
   "repository": "github:saltcorn/saltcorn",
   "publishConfig": {
     "access": "public"
+  },
+  "devDependencies": {
+    "@types/node": "^16.11.9",
+    "@types/sqlite3": "^3.1.7",
+    "typescript": "^4.4.4"
   }
 }
