+++ packages/sqlite/src/index.ts
@@ -11,4 +11,3 @@
  *
  * @category sqlite
  */
-
