+++ packages/sqlite/src/sqlite.ts
@@ -1,0 +1,312 @@
+/**
+ * SQLite3 data access layer
+ * @category sqlite
+ * @module sqlite
+ */
+
+import { Database, verbose } from "sqlite3";
+verbose();
+import { unlink } from "fs/promises"
+
+const { sqlsanitize, mkWhere, mkSelectOptions } = require("@saltcorn/db-common/internal");
+
+
+let sqliteDatabase:Database|null = null;
+let connectObj:any = null;
+let current_filepath:string;
+
+/**
+ * Initializes internals of the the sqlite module.
+ * It must be called after importing the module.
+ * @param {any} getConnectObject 
+ */
+export const init = (getConnectObject: () => Database): void => {
+  if (!sqliteDatabase) {
+    connectObj = getConnectObject();
+    current_filepath = get_db_filepath();
+    sqliteDatabase = new Database(current_filepath);
+  }
+}
+
+/**
+ * Get sqlite path
+ * @returns {string|void}
+ */
+export const get_db_filepath = (): string => {
+  if (connectObj.sqlite_path) return connectObj.sqlite_path;
+  return "";
+};
+
+let log_sql_enabled = false;
+
+/**
+ * Control Logging sql statements to console
+ * @param {boolean} [val = true] - if true then log sql statements to console
+ */
+export function set_sql_logging(val:boolean = true): void {
+  log_sql_enabled = val;
+}
+
+/**
+ * Get sql logging state
+ * @returns {boolean} if true then sql logging eabled
+ */
+export function get_sql_logging(): boolean {
+  return log_sql_enabled;
+}
+
+/**
+ * Log SQL statement to console
+ * @param {string} sql - SQL statement
+ * @param {any} [vs] - any additional parameter
+ */
+export function sql_log(sql:string, vs?:any) {
+  if (log_sql_enabled)
+    if (typeof vs === "undefined") console.log(sql);
+    else console.log(sql, vs);
+}
+
+/**
+ * @param {string} sql 
+ * @param {any} params 
+ * @returns {Promise<object>}
+ */
+export function query(sql:string, params?:any): Promise<any> {
+  sql_log(sql, params);
+  return new Promise((resolve, reject) => {
+    if(!sqliteDatabase) {
+      throw new Error("The database connection is closed.");
+    }
+    sqliteDatabase.all(sql, params, function (err:any, rows:any) {
+      if (err) {
+        reject(err);
+      } else {
+        resolve({ rows });
+      }
+    });
+  });
+}
+
+/**
+ * Change connection (close connection and open new connection from connObj)
+ * @param {any} connObj - connection object
+ * @returns {Promise<void>}
+ */
+export const changeConnection = async (connObj:any): Promise<void> => {
+  if(!sqliteDatabase) {
+    throw new Error("The database connection is closed.");
+  }
+  await sqliteDatabase.close();
+  current_filepath = connObj.sqlite_path;
+  sqliteDatabase = new Database(current_filepath);
+};
+
+/**
+ * Close database connection
+ * @returns {Promise<void>}
+ */
+export const close = async (): Promise<void> => {
+  if(!sqliteDatabase) {
+    throw new Error("The database connection is closed.");
+  }
+  await sqliteDatabase.close();
+  sqliteDatabase = null;
+};
+
+/**
+ * Execute Select statement
+ * @param {string} tbl - table name
+ * @param {any} whereObj - where object
+ * @param {any} [selectopts = {}] - select options
+ * @returns {Promise<*>} return rows
+ */
+export const select = async (tbl:string, whereObj:any, selectopts:any = {})
+: Promise<any> => {
+  const { where, values } = mkWhere(whereObj, true);
+  const sql = `SELECT * FROM "${sqlsanitize(tbl)}" ${where} ${mkSelectOptions(
+    selectopts
+  )}`;
+  const tq = await query(sql, values);
+
+  return tq.rows;
+};
+
+
+// TODO Utility function - needs ti be moved out this module
+/**
+ * @param {any} v 
+ * @returns {boolean}
+ */
+export const reprAsJson = (v:any): boolean =>
+  typeof v === "object" && v !== null && !(v instanceof Date);
+
+/**
+ * @param {object[]} opts
+ * @param {*} opts.k
+ * @param {any} opts.v
+ * @returns {string|any}
+ */
+export const mkVal = ([k, v]:[any, any]): string => (reprAsJson(v) ? JSON.stringify(v) : v);
+
+/**
+ * Drop unique constraint
+ * @param {string} tbl - table name
+ * @param {any} obj - list of column=value pairs
+ * @param {string} id - primary key column value
+ * @returns {Promise<void>} no results
+ */
+export const update = async (tbl:string, obj:any, id:string): Promise<void>  => {
+  const kvs = Object.entries(obj);
+  const assigns = kvs.map(([k, v], ix) => `"${sqlsanitize(k)}"=?`).join();
+  let valList = kvs.map(mkVal);
+  valList.push(id);
+  const q = `update "${sqlsanitize(tbl)}" set ${assigns} where id=?`;
+  await query(q, valList);
+};
+
+/**
+ * Delete rows in table
+ * @param {string} tbl - table name
+ * @param {any} whereObj - where object
+ * @returns {Promise<void>} result of delete execution
+ */
+export const deleteWhere = async (tbl:string, whereObj:any): Promise<void> => {
+  const { where, values } = mkWhere(whereObj, true);
+  const sql = `delete FROM "${sqlsanitize(tbl)}" ${where}`;
+
+  const tq = await query(sql, values);
+};
+
+/**
+ * Insert rows into table
+ * @param {string} tbl - table name
+ * @param {any} obj - columns names and data
+ * @param {any} [opts = {}] - columns attributes
+ * @returns {Promise<string|void>} returns id.
+ */
+export const insert = async (tbl:string, obj:any, opts:any = {}): Promise<string|void> => {
+  const kvs:any = Object.entries(obj);
+  const fnameList = kvs.map(([k, v]:[any, any]) => `"${sqlsanitize(k)}"`).join();
+  const valPosList = kvs
+    .map(([k, v]:[any, any], ix:any) =>
+      v && v.next_version_by_id
+        ? `coalesce((select max(_version) from "${sqlsanitize(
+            tbl
+          )}" where id=${+v.next_version_by_id}), 0)+1`
+        : reprAsJson(v)
+        ? "json(?)"
+        : "?"
+    )
+    .join();
+  const valList = kvs
+    .filter(([k, v]:[any, any]) => !(v && v.next_version_by_id))
+    .map(mkVal);
+  const sql = `insert into "${sqlsanitize(
+    tbl
+  )}"(${fnameList}) values(${valPosList})`;
+
+  await query(sql, valList);
+  if (opts.noid) return;
+  // TBD Support of primary key column different from id
+  const ids = await query("SELECT last_insert_rowid() as id");
+  return ids.rows[0].id;
+};
+
+/**
+ * Select one record
+ * @param {string} tbl - table name
+ * @param {any} where - where object
+ * @throws {Error}
+ * @returns {Promise<any>} return first record from sql result
+ */
+export const selectOne = async (tbl:string, where:any): Promise<any> => {
+  const rows = await select(tbl, where);
+  if (rows.length === 0) {
+    const w = mkWhere(where, true);
+    throw new Error(`no ${tbl} ${w.where} are ${w.values}`);
+  } else return rows[0];
+};
+
+/**
+ * Select one record or null if no records
+ * @param {string} tbl - table name
+ * @param {any} where - where object
+ * @returns {Promise<any>} - null if no record or first record data
+ */
+export const selectMaybeOne = async (tbl:string, where:any): Promise<any> => {
+  const rows = await select(tbl, where);
+  if (rows.length === 0) return null;
+  else return rows[0];
+};
+
+/**
+ * Get count of rows in table
+ * @param {string} tbl - table name
+ * @param {any} whereObj - where object
+ * @returns {Promise<number>} count of tables
+ */
+export const count = async (tbl:string, whereObj:any) => {
+  const { where, values } = mkWhere(whereObj, true);
+  const sql = `SELECT COUNT(*) FROM "${sqlsanitize(tbl)}" ${where}`;
+  const tq = await query(sql, values);
+  return parseInt(tq.rows[0]["COUNT(*)"]);
+};
+
+/**
+ * Get version of PostgreSQL
+ * @returns {Promise<string>} returns version
+ */
+export const getVersion = async (): Promise<string> => {
+  const sql = `SELECT sqlite_version();`;
+  sql_log(sql);
+  const tq = await query(sql);
+  return tq.rows[0]["sqlite_version()"];
+};
+
+/**
+ * Reset DB Schema using drop schema and recreate it.
+ * Attention! You will lost data after call this function!
+ * @returns {Promise<void>} no result
+ */
+export const drop_reset_schema = async (): Promise<void> => {
+  if(!sqliteDatabase) {
+    throw new Error("The database connection is closed.");
+  }
+  await sqliteDatabase.close();
+  await unlink(current_filepath);
+  sqliteDatabase = new Database(current_filepath);
+};
+
+/**
+ * Add unique constraint
+ * @param {string} table_name - table name
+ * @param {string[]} field_names - list of columns (members of constraint)
+ * @returns {Promise<void>} no result
+ */
+export const add_unique_constraint = async (table_name:string, field_names:string[])
+: Promise<void> => {
+  const sql = `create unique index ${sqlsanitize(
+    table_name
+  )}_${field_names
+    .map((f) => sqlsanitize(f))
+    .join("_")}_unique on "${sqlsanitize(table_name)}"(${field_names
+    .map((f) => `"${sqlsanitize(f)}"`)
+    .join(",")});`;
+  sql_log(sql);
+  await query(sql);
+};
+
+/**
+ * Drop unique constraint
+ * @param {string} table_name - table name
+ * @param {string[]} field_names - list of columns (members of constraint)
+ * @returns {Promise<void>} no results
+ */
+export const drop_unique_constraint = async (table_name:string, field_names:string[])
+: Promise<void> => {
+  const sql = `drop index ${sqlsanitize(table_name)}_${field_names
+    .map((f) => sqlsanitize(f))
+    .join("_")}_unique;`;
+  sql_log(sql);
+  await query(sql);
+};
