+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "files": [],
+  "references": [
+    {
+      "path": "./packages/sqlite/tsconfig.ref.json"
+    }
+  ],
+}
